requirements.txt
Django
xmpppy
